Changed photos to display at 100% width in a feed format - #3
We'll probably want to make all the photos a universal height as the
landscape pictures can be longer than the entire screen. I'm not sure
how it is on android but iOS doesn't appear to do well for even small
thumbnails grouped 20 images at a time.

This appears to be the only way I'm getting any type of performance from
iOS.